commit of shame. Fixed bug: DEREFERENCING A NULLPTR. Did other stuffs too

+++ FakeAPlib/src/include/SDCardManager.h
@@ -8,7 +8,8 @@ class SDCardManager
 {
 protected:
     fs::FS _fileSystem;
-    bool _sdInitialized;
+    bool   _sdInitialized;
+
 
 public:
     SDCardManager()
@@ -16,12 +17,26 @@ public:
     {
     }
 
+    SDCardManager(const SDCardManager&) = delete;
+
+    SDCardManager operator= (const SDCardManager&) = delete;
+
+    void operator+ (const String& path)
+    {
+        if (!writeFile(path.c_str())) return;
+    }
+
+    void operator- (const String& path)
+    {
+        if (!deleteFile(path.c_str())) return;
+    }
+
     fs::FS& getFileSystem()
     {
         return _fileSystem;
     }
 
-    char *getCardType(uint8_t card)
+    static char *getCardType(uint8_t card)
     {
         switch (card)
         {
@@ -42,15 +57,28 @@ public:
 
     inline bool logEvent(const String &event)
     {
-        File logFile = SD.open(LOG_FILE, FILE_APPEND);
-        if (!logFile)
+        if (!_sdInitialized)
         {
-            if (WITH_ERROR_TYPE) Serial.printf(ERROR_FILE_OPEN, logFile);
+            if (WITH_ERROR_TYPE) Serial.println(ERROR_SD_NOT_INIT);
             return 0;
         }
 
-        logFile.println(String(millis()) + ": " + event);
+        if(!_fileSystem.exists(LOG_FILE))
+        {
+            //writeFile(LOG_FILE);
+            return writeFile(LOG_FILE, getCurrentTime() + ": " + event)? 1 : 0;
+        }
+
+        File logFile = _fileSystem.open(LOG_FILE, FILE_APPEND);
+        if (!logFile)
+        {
+            if (WITH_ERROR_TYPE) Serial.printf(ERROR_FILE_OPEN, LOG_FILE);
+            return 0;
+        }
+
+        logFile.println(getCurrentTime() + ": " + event);
         logFile.close();
+        Serial.println(SUCCESS_LOG_EVENT);
 
         return 1;
     }
@@ -73,46 +101,79 @@ public:
             return 0;
         }
 
-        Serial.println("Tarjeta SD inicializada correctamente.");
+        Serial.println(SUCCESS_SD_INIT);
         Serial.printf("Tipo de tarjeta: %s\n", getCardType(cardType));
 
         uint64_t cardSize = SD_MMC.cardSize() / (1024 * 1024);
         Serial.printf("Tamaño de la tarjeta SD_MMC: %lluMB\n", cardSize);
 
-        listDir("/", 0);
+        _sdInitialized = true;
 
         Serial.printf("Total espacio disponible: %lluMB\r\n", SD_MMC.totalBytes() / (1024 * 1024));
 
-        _sdInitialized = true;
+        listDir("/", 0);
 
         return 1;
     }
 
     inline String readFile(const char *path)
     {
+        if (!_sdInitialized)
+        {
+            if (WITH_ERROR_TYPE) Serial.println(ERROR_SD_NOT_INIT);
+            return "\0";
+        }
+
         File file = _fileSystem.open(path);
         if (!file)
         {
             if (WITH_ERROR_TYPE) Serial.printf(ERROR_FILE_OPEN, path);
             return "\0";
         }
-        else if (!_sdInitialized)
+
+        String content = "";
+        while (file.available())
+            content += (char)file.read();
+        
+        Serial.printf(SUCCESS_FILE_OPEN, file.name());
+        file.close();
+
+        return content;
+    }
+
+    inline String readFile(const String& path)
+    {
+        if (!_sdInitialized)
         {
             if (WITH_ERROR_TYPE) Serial.println(ERROR_SD_NOT_INIT);
+            return "\0";
+        }
+
+        File file = _fileSystem.open(path);
+        if (!file)
+        {
+            if (WITH_ERROR_TYPE) Serial.printf(ERROR_FILE_OPEN, path.c_str());
             return "\0";
         }
 
         String content = "";
         while (file.available())
             content += (char)file.read();
-
+        
+        Serial.printf(SUCCESS_FILE_OPEN, file.name());
         file.close();
 
         return content;
     }
 
-    inline bool writeFile(const char *path, String &content)
+    inline bool writeFile(const char *path, const String &content="")
     {
+        if (!_sdInitialized)
+        {
+            if (WITH_ERROR_TYPE) Serial.println(ERROR_SD_NOT_INIT);
+            return 0;
+        }
+        
         File file = _fileSystem.open(path, FILE_WRITE);
         if (!file)
         {
@@ -124,14 +185,44 @@ public:
             if (WITH_ERROR_TYPE) Serial.printf(ERROR_FILE_WRITE, path);
             return 0;
         }
+        Serial.printf(SUCCESS_FILE_WRITE, file.name());
+        file.close();
+        return 1;
+    }
+
+    inline bool writeFile(const String& path, const String &content="")
+    {
+        if (!_sdInitialized)
+        {
+            if (WITH_ERROR_TYPE) Serial.println(ERROR_SD_NOT_INIT);
+            return 0;
+        }
+        
+        File file = _fileSystem.open(path, FILE_WRITE);
+        if (!file)
+        {
+            if (WITH_ERROR_TYPE) Serial.printf(ERROR_FILE_OPEN, path.c_str());
+            return 0;
+        }
+        else if (!file.println(content))
+        {
+            if (WITH_ERROR_TYPE) Serial.printf(ERROR_FILE_WRITE, path.c_str());
+            return 0;
+        }
+        Serial.printf(SUCCESS_FILE_WRITE, file.name());
         file.close();
         return 1;
     }
 
     inline void listDir(const char *dirname, uint8_t levels)
     {
-        //Serial.printf("Listing directory: %s\n", dirname);
+        if (!_sdInitialized)
+        {
+            if (WITH_ERROR_TYPE) Serial.println(ERROR_SD_NOT_INIT);
+            return;
+        }
 
+        Serial.printf("Listing directory: %s\n", dirname);
         File root = _fileSystem.open(dirname);
         if (!root)
         {
@@ -168,29 +259,49 @@ public:
 
     inline bool createDir(const char *path)
     {
+        if (!_sdInitialized)
+        {
+            if (WITH_ERROR_TYPE) Serial.println(ERROR_SD_NOT_INIT);
+            return 0;
+        }
+
         Serial.printf("Creating Dir: %s\n", path);
         if (!_fileSystem.mkdir(path))
         {
             if (WITH_ERROR_TYPE) Serial.printf(ERROR_DIR_CREATE, path);
             return 0;
         }
-
+        
+        Serial.printf(SUCCESS_DIR_CREATE, path);
         return 1;
     }
 
     inline bool removeDir(const char *path)
     {
+        if (!_sdInitialized)
+        {
+            if (WITH_ERROR_TYPE) Serial.println(ERROR_SD_NOT_INIT);
+            return 0;
+        }
+
         Serial.printf("Removing Dir: %s\n", path);
         if (!_fileSystem.rmdir(path))
         {
             if (WITH_ERROR_TYPE) Serial.printf(ERROR_DIR_DELETE, path);
             return 0;
         }
+
+        Serial.printf(SUCCESS_DIR_DELETE, path);
         return 1;
     }
 
     inline bool appendFile(const char *path, String &content)
     {
+        if (!_sdInitialized)
+        {
+            if (WITH_ERROR_TYPE) Serial.println(ERROR_SD_NOT_INIT);
+            return 0;
+        }
         File file = _fileSystem.open(path, FILE_APPEND);
         if (!file)
         {
@@ -203,6 +314,32 @@ public:
             Serial.printf(ERROR_FILE_WRITE, path);
             return 0;
         }
+        Serial.printf(SUCCESS_FILE_WRITE, file.name());
+        file.close();
+
+        return 1;
+    }
+
+    inline bool appendFile(const String& path, String &content)
+    {
+        if (!_sdInitialized)
+        {
+            if (WITH_ERROR_TYPE) Serial.println(ERROR_SD_NOT_INIT);
+            return 0;
+        }
+        File file = _fileSystem.open(path, FILE_APPEND);
+        if (!file)
+        {
+            if (WITH_ERROR_TYPE) Serial.printf(ERROR_FILE_OPEN, path.c_str());
+            return 0;
+        }
+        content = "\n" + content;
+        if (!file.print(content))
+        {
+            Serial.printf(ERROR_FILE_WRITE, path.c_str());
+            return 0;
+        }
+        Serial.printf(SUCCESS_FILE_WRITE, file.name());
         file.close();
 
         return 1;
@@ -210,29 +347,49 @@ public:
 
     inline bool renameFile(const char *path1, const char *path2)
     {
+        if (!_sdInitialized)
+        {
+            if (WITH_ERROR_TYPE) Serial.println(ERROR_SD_NOT_INIT);
+            return 0;
+        }
+
         if (!_fileSystem.rename(path1, path2))
         {
             if (WITH_ERROR_TYPE) Serial.printf(ERROR_FILE_RENAME, path1);
             return 0;
         }
+        Serial.printf(SUCCESS_FILE_RENAME, path1);
 
         return 1;
     }
 
     inline bool deleteFile(const char *path)
     {
+        if (!_sdInitialized)
+        {
+            if (WITH_ERROR_TYPE) Serial.println(ERROR_SD_NOT_INIT);
+            return 0;
+        }
+
         Serial.printf("Deleting file: %s\n", path);
         if (!_fileSystem.remove(path))
         {
             if (WITH_ERROR_TYPE) Serial.printf(ERROR_FILE_DELETE, path);
             return 0;
         }
+        Serial.printf(SUCCESS_FILE_DELETE, path);
 
         return 1;
     }
 
     inline void testFileIO(const char *path)
     {
+        if (!_sdInitialized)
+        {
+            if (WITH_ERROR_TYPE) Serial.println(ERROR_SD_NOT_INIT);
+            return;
+        }
+
         File file = _fileSystem.open(path);
         static uint8_t buf[512];
         size_t len = 0;
@@ -279,6 +436,8 @@ public:
         Serial.printf("%u bytes written for %u ms\n", 2048 * 512, end);
         file.close();
     }
+
+    inline bool isCardInit () const { return _sdInitialized; }
 };
 
 #endif
